FTM004-CustomCell: Added CWStatusNotification Support and a TMTravelDetailsController delegate.
Added CWStatusNotification support to alert user when a Travel marker has been added to the map or when the travel choice has been discarded. Added a TMTravelDetailsController  delegate to help present a "Added new Travel" CWStatusNotification when a Travel is stored and to also handle dropping the markers with their animation after the TMTravelDetailsController is dismissed. Added a UIImageView to the TMTravelDetailsController view. Did UI Work. Documented new methods. Moved stuff around.